+++ components/compare-frames/cache-handler.js
