+++ components/compare-frames/image-processor.js
