+++ components/compare-frames/index.js
