+++ components/compare-frames/utils.js
