+++ components/extract-frames.js
