+++ sequencers/diff-wav-rider.js
@@ -1,4 +1,4 @@
-const seq = require("../src/sequitur");
+const seq = require("..");
 
 (async () => {
   const wave = seq.wave();
