+++ sequencers/pathfinder.js
@@ -1,4 +1,4 @@
-const seq = require("../src/sequitur");
+const seq = require("..");
 
 (async () => {
   const diffs = await seq.diffs(undefined, undefined, true);
